openfimg: compiler: Enforce non-temporary register access constraint
Non-temporary registers can be accessed only once per instruction per
register type, which means that for an instruction that has more than
one operand of particular non-temporary type, all but one of them must
be preloaded to temporary registers.

The pass doing the enforcement is very naive, inserting movs for
subsequent operands of given type after first, so there is room for
improvement.

Signed-off-by: Tomasz Figa <tomasz.figa@gmail.com>