openfimg: compiler: Add handling of register parity in 3-source ops
Each instruction can access registers of the same parity at most two
times, which means that there might not be three temporary operands of
the same parity in a 3-source instruction.

The code is still a bit too conservative as it simply semi-randomly
assigns parity constraint for each operand of 3-source instructions in
a globally round robin fashion. However in practice it should work fine.
The only case current code fails is obviously when all three operands
are the same variables, which is not very likely to be spotted in real
shaders, but needs to be fixed anyway.

Signed-off-by: Tomasz Figa <tomasz.figa@gmail.com>