openfimg: compiler: Work around pixel shader termination on output write
Apparently the shader ends execution as soon as a write to the output
color register happens, even if it writes only selected components. To
work this around, a temporary register is used throughout the program,
which is then stored to the output register at the end.

Possible valid cases of writing directly to the output register will be
handled later by respective pass in the optimizer.

Signed-off-by: Tomasz Figa <tomasz.figa@gmail.com>